build

+++ src/pages/abi-decode/AbiDecode.sol
@@ -4,26 +4,30 @@ pragma solidity ^0.8.20;
 contract AbiDecode {
     struct MyStruct {
         string name;
-        uint[2] nums;
+        uint256[2] nums;
     }
 
     function encode(
-        uint x,
+        uint256 x,
         address addr,
-        uint[] calldata arr,
+        uint256[] calldata arr,
         MyStruct calldata myStruct
     ) external pure returns (bytes memory) {
         return abi.encode(x, addr, arr, myStruct);
     }
 
-    function decode(
-        bytes calldata data
-    )
+    function decode(bytes calldata data)
         external
         pure
-        returns (uint x, address addr, uint[] memory arr, MyStruct memory myStruct)
+        returns (
+            uint256 x,
+            address addr,
+            uint256[] memory arr,
+            MyStruct memory myStruct
+        )
     {
         // (uint x, address addr, uint[] memory arr, MyStruct myStruct) = ...
-        (x, addr, arr, myStruct) = abi.decode(data, (uint, address, uint[], MyStruct));
+        (x, addr, arr, myStruct) =
+            abi.decode(data, (uint256, address, uint256[], MyStruct));
     }
 }
