+++ src/pages/app/bi-directional-payment-channel/BiDirectionalPaymentChannel.sol
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
+import "./ECDSA.sol";
+
 /*
 Opening a channel
 1. Alice and Bob fund a multi-sig wallet
@@ -27,24 +29,22 @@ Closing a channel when Alice and Bob do not agree on the final balances
 3. Alice and Bob can withdraw funds once the channel is expired
 */
 
-import "github.com/OpenZeppelin/openzeppelin-contracts/blob/release-v4.5/contracts/utils/cryptography/ECDSA.sol";
-
 contract BiDirectionalPaymentChannel {
     using ECDSA for bytes32;
 
-    event ChallengeExit(address indexed sender, uint nonce);
-    event Withdraw(address indexed to, uint amount);
+    event ChallengeExit(address indexed sender, uint256 nonce);
+    event Withdraw(address indexed to, uint256 amount);
 
     address payable[2] public users;
     mapping(address => bool) public isUser;
 
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
-    uint public challengePeriod;
-    uint public expiresAt;
-    uint public nonce;
+    uint256 public challengePeriod;
+    uint256 public expiresAt;
+    uint256 public nonce;
 
-    modifier checkBalances(uint[2] memory _balances) {
+    modifier checkBalances(uint256[2] memory _balances) {
         require(
             address(this).balance >= _balances[0] + _balances[1],
             "balance of contract must be >= to the total balance of users"
@@ -55,14 +55,14 @@ contract BiDirectionalPaymentChannel {
     // NOTE: deposit from multi-sig wallet
     constructor(
         address payable[2] memory _users,
-        uint[2] memory _balances,
-        uint _expiresAt,
-        uint _challengePeriod
+        uint256[2] memory _balances,
+        uint256 _expiresAt,
+        uint256 _challengePeriod
     ) payable checkBalances(_balances) {
         require(_expiresAt > block.timestamp, "Expiration must be > now");
         require(_challengePeriod > 0, "Challenge period must be > 0");
 
-        for (uint i = 0; i < _users.length; i++) {
+        for (uint256 i = 0; i < _users.length; i++) {
             address payable user = _users[i];
 
             require(!isUser[user], "user must be unique");
@@ -80,18 +80,17 @@ contract BiDirectionalPaymentChannel {
         bytes[2] memory _signatures,
         address _contract,
         address[2] memory _signers,
-        uint[2] memory _balances,
-        uint _nonce
+        uint256[2] memory _balances,
+        uint256 _nonce
     ) public pure returns (bool) {
-        for (uint i = 0; i < _signatures.length; i++) {
+        for (uint256 i = 0; i < _signatures.length; i++) {
             /*
             NOTE: sign with address of this contract to protect
                   agains replay attack on other contracts
             */
-            bool valid = _signers[i] ==
-                keccak256(abi.encodePacked(_contract, _balances, _nonce))
-                    .toEthSignedMessageHash()
-                    .recover(_signatures[i]);
+            bool valid = _signers[i]
+                == keccak256(abi.encodePacked(_contract, _balances, _nonce))
+                    .toEthSignedMessageHash().recover(_signatures[i]);
 
             if (!valid) {
                 return false;
@@ -103,12 +102,12 @@ contract BiDirectionalPaymentChannel {
 
     modifier checkSignatures(
         bytes[2] memory _signatures,
-        uint[2] memory _balances,
-        uint _nonce
+        uint256[2] memory _balances,
+        uint256 _nonce
     ) {
         // Note: copy storage array to memory
         address[2] memory signers;
-        for (uint i = 0; i < users.length; i++) {
+        for (uint256 i = 0; i < users.length; i++) {
             signers[i] = users[i];
         }
 
@@ -126,8 +125,8 @@ contract BiDirectionalPaymentChannel {
     }
 
     function challengeExit(
-        uint[2] memory _balances,
-        uint _nonce,
+        uint256[2] memory _balances,
+        uint256 _nonce,
         bytes[2] memory _signatures
     )
         public
@@ -138,7 +137,7 @@ contract BiDirectionalPaymentChannel {
         require(block.timestamp < expiresAt, "Expired challenge period");
         require(_nonce > nonce, "Nonce must be greater than the current nonce");
 
-        for (uint i = 0; i < _balances.length; i++) {
+        for (uint256 i = 0; i < _balances.length; i++) {
             balances[users[i]] = _balances[i];
         }
 
@@ -149,12 +148,14 @@ contract BiDirectionalPaymentChannel {
     }
 
     function withdraw() public onlyUser {
-        require(block.timestamp >= expiresAt, "Challenge period has not expired yet");
+        require(
+            block.timestamp >= expiresAt, "Challenge period has not expired yet"
+        );
 
-        uint amount = balances[msg.sender];
+        uint256 amount = balances[msg.sender];
         balances[msg.sender] = 0;
 
-        (bool sent, ) = msg.sender.call{value: amount}("");
+        (bool sent,) = msg.sender.call{value: amount}("");
         require(sent, "Failed to send Ether");
 
         emit Withdraw(msg.sender, amount);
