+++ src/pages/app/create2/Create2.sol
@@ -3,11 +3,11 @@ pragma solidity ^0.8.20;
 
 contract Factory {
     // Returns the address of the newly deployed contract
-    function deploy(
-        address _owner,
-        uint _foo,
-        bytes32 _salt
-    ) public payable returns (address) {
+    function deploy(address _owner, uint256 _foo, bytes32 _salt)
+        public
+        payable
+        returns (address)
+    {
         // This syntax is a newer way to invoke create2 without assembly, you just need to pass salt
         // https://docs.soliditylang.org/en/latest/control-structures.html#salted-contract-creations-create2
         return address(new TestContract{salt: _salt}(_owner, _foo));
@@ -16,11 +16,15 @@ contract Factory {
 
 // This is the older way of doing it using assembly
 contract FactoryAssembly {
-    event Deployed(address addr, uint salt);
+    event Deployed(address addr, uint256 salt);
 
     // 1. Get bytecode of contract to be deployed
     // NOTE: _owner and _foo are arguments of the TestContract's constructor
-    function getBytecode(address _owner, uint _foo) public pure returns (bytes memory) {
+    function getBytecode(address _owner, uint256 _foo)
+        public
+        pure
+        returns (bytes memory)
+    {
         bytes memory bytecode = type(TestContract).creationCode;
 
         return abi.encodePacked(bytecode, abi.encode(_owner, _foo));
@@ -28,23 +32,26 @@ contract FactoryAssembly {
 
     // 2. Compute the address of the contract to be deployed
     // NOTE: _salt is a random number used to create an address
-    function getAddress(
-        bytes memory bytecode,
-        uint _salt
-    ) public view returns (address) {
+    function getAddress(bytes memory bytecode, uint256 _salt)
+        public
+        view
+        returns (address)
+    {
         bytes32 hash = keccak256(
-            abi.encodePacked(bytes1(0xff), address(this), _salt, keccak256(bytecode))
+            abi.encodePacked(
+                bytes1(0xff), address(this), _salt, keccak256(bytecode)
+            )
         );
 
         // NOTE: cast last 20 bytes of hash to address
-        return address(uint160(uint(hash)));
+        return address(uint160(uint256(hash)));
     }
 
     // 3. Deploy the contract
     // NOTE:
     // Check the event log Deployed which contains the address of the deployed TestContract.
     // The address in the log should equal the address computed from above.
-    function deploy(bytes memory bytecode, uint _salt) public payable {
+    function deploy(bytes memory bytecode, uint256 _salt) public payable {
         address addr;
 
         /*
@@ -58,17 +65,16 @@ contract FactoryAssembly {
               s = big-endian 256-bit value
         */
         assembly {
-            addr := create2(
-                callvalue(), // wei sent with current call
-                // Actual code starts after skipping the first 32 bytes
-                add(bytecode, 0x20),
-                mload(bytecode), // Load the size of code contained in the first 32 bytes
-                _salt // Salt from function arguments
-            )
+            addr :=
+                create2(
+                    callvalue(), // wei sent with current call
+                    // Actual code starts after skipping the first 32 bytes
+                    add(bytecode, 0x20),
+                    mload(bytecode), // Load the size of code contained in the first 32 bytes
+                    _salt // Salt from function arguments
+                )
 
-            if iszero(extcodesize(addr)) {
-                revert(0, 0)
-            }
+            if iszero(extcodesize(addr)) { revert(0, 0) }
         }
 
         emit Deployed(addr, _salt);
@@ -77,14 +83,14 @@ contract FactoryAssembly {
 
 contract TestContract {
     address public owner;
-    uint public foo;
+    uint256 public foo;
 
-    constructor(address _owner, uint _foo) payable {
+    constructor(address _owner, uint256 _foo) payable {
         owner = _owner;
         foo = _foo;
     }
 
-    function getBalance() public view returns (uint) {
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
