+++ src/pages/app/crowd-fund/CrowdFund.sol
@@ -2,32 +2,31 @@
 pragma solidity ^0.8.20;
 
 interface IERC20 {
-    function transfer(address, uint) external returns (bool);
-
-    function transferFrom(address, address, uint) external returns (bool);
+    function transfer(address, uint256) external returns (bool);
+    function transferFrom(address, address, uint256) external returns (bool);
 }
 
 contract CrowdFund {
     event Launch(
-        uint id,
+        uint256 id,
         address indexed creator,
-        uint goal,
+        uint256 goal,
         uint32 startAt,
         uint32 endAt
     );
-    event Cancel(uint id);
-    event Pledge(uint indexed id, address indexed caller, uint amount);
-    event Unpledge(uint indexed id, address indexed caller, uint amount);
-    event Claim(uint id);
-    event Refund(uint id, address indexed caller, uint amount);
+    event Cancel(uint256 id);
+    event Pledge(uint256 indexed id, address indexed caller, uint256 amount);
+    event Unpledge(uint256 indexed id, address indexed caller, uint256 amount);
+    event Claim(uint256 id);
+    event Refund(uint256 id, address indexed caller, uint256 amount);
 
     struct Campaign {
         // Creator of campaign
         address creator;
         // Amount of tokens to raise
-        uint goal;
+        uint256 goal;
         // Total amount pledged
-        uint pledged;
+        uint256 pledged;
         // Timestamp of start of campaign
         uint32 startAt;
         // Timestamp of end of campaign
@@ -39,17 +38,17 @@ contract CrowdFund {
     IERC20 public immutable token;
     // Total count of campaigns created.
     // It is also used to generate id for new campaigns.
-    uint public count;
+    uint256 public count;
     // Mapping from id to Campaign
-    mapping(uint => Campaign) public campaigns;
+    mapping(uint256 => Campaign) public campaigns;
     // Mapping from campaign id => pledger => amount pledged
-    mapping(uint => mapping(address => uint)) public pledgedAmount;
+    mapping(uint256 => mapping(address => uint256)) public pledgedAmount;
 
     constructor(address _token) {
         token = IERC20(_token);
     }
 
-    function launch(uint _goal, uint32 _startAt, uint32 _endAt) external {
+    function launch(uint256 _goal, uint32 _startAt, uint32 _endAt) external {
         require(_startAt >= block.timestamp, "start at < now");
         require(_endAt >= _startAt, "end at < start at");
         require(_endAt <= block.timestamp + 90 days, "end at > max duration");
@@ -67,7 +66,7 @@ contract CrowdFund {
         emit Launch(count, msg.sender, _goal, _startAt, _endAt);
     }
 
-    function cancel(uint _id) external {
+    function cancel(uint256 _id) external {
         Campaign memory campaign = campaigns[_id];
         require(campaign.creator == msg.sender, "not creator");
         require(block.timestamp < campaign.startAt, "started");
@@ -76,7 +75,7 @@ contract CrowdFund {
         emit Cancel(_id);
     }
 
-    function pledge(uint _id, uint _amount) external {
+    function pledge(uint256 _id, uint256 _amount) external {
         Campaign storage campaign = campaigns[_id];
         require(block.timestamp >= campaign.startAt, "not started");
         require(block.timestamp <= campaign.endAt, "ended");
@@ -88,7 +87,7 @@ contract CrowdFund {
         emit Pledge(_id, msg.sender, _amount);
     }
 
-    function unpledge(uint _id, uint _amount) external {
+    function unpledge(uint256 _id, uint256 _amount) external {
         Campaign storage campaign = campaigns[_id];
         require(block.timestamp <= campaign.endAt, "ended");
 
@@ -99,7 +98,7 @@ contract CrowdFund {
         emit Unpledge(_id, msg.sender, _amount);
     }
 
-    function claim(uint _id) external {
+    function claim(uint256 _id) external {
         Campaign storage campaign = campaigns[_id];
         require(campaign.creator == msg.sender, "not creator");
         require(block.timestamp > campaign.endAt, "not ended");
@@ -112,12 +111,12 @@ contract CrowdFund {
         emit Claim(_id);
     }
 
-    function refund(uint _id) external {
+    function refund(uint256 _id) external {
         Campaign memory campaign = campaigns[_id];
         require(block.timestamp > campaign.endAt, "not ended");
         require(campaign.pledged < campaign.goal, "pledged >= goal");
 
-        uint bal = pledgedAmount[_id][msg.sender];
+        uint256 bal = pledgedAmount[_id][msg.sender];
         pledgedAmount[_id][msg.sender] = 0;
         token.transfer(msg.sender, bal);
