+++ src/pages/app/dutch-auction/DutchAuction.sol
@@ -2,48 +2,56 @@
 pragma solidity ^0.8.20;
 
 interface IERC721 {
-    function transferFrom(address _from, address _to, uint _nftId) external;
+    function transferFrom(address _from, address _to, uint256 _nftId)
+        external;
 }
 
 contract DutchAuction {
-    uint private constant DURATION = 7 days;
+    uint256 private constant DURATION = 7 days;
 
     IERC721 public immutable nft;
-    uint public immutable nftId;
+    uint256 public immutable nftId;
 
     address payable public immutable seller;
-    uint public immutable startingPrice;
-    uint public immutable startAt;
-    uint public immutable expiresAt;
-    uint public immutable discountRate;
+    uint256 public immutable startingPrice;
+    uint256 public immutable startAt;
+    uint256 public immutable expiresAt;
+    uint256 public immutable discountRate;
 
-    constructor(uint _startingPrice, uint _discountRate, address _nft, uint _nftId) {
+    constructor(
+        uint256 _startingPrice,
+        uint256 _discountRate,
+        address _nft,
+        uint256 _nftId
+    ) {
         seller = payable(msg.sender);
         startingPrice = _startingPrice;
         startAt = block.timestamp;
         expiresAt = block.timestamp + DURATION;
         discountRate = _discountRate;
 
-        require(_startingPrice >= _discountRate * DURATION, "starting price < min");
+        require(
+            _startingPrice >= _discountRate * DURATION, "starting price < min"
+        );
 
         nft = IERC721(_nft);
         nftId = _nftId;
     }
 
-    function getPrice() public view returns (uint) {
-        uint timeElapsed = block.timestamp - startAt;
-        uint discount = discountRate * timeElapsed;
+    function getPrice() public view returns (uint256) {
+        uint256 timeElapsed = block.timestamp - startAt;
+        uint256 discount = discountRate * timeElapsed;
         return startingPrice - discount;
     }
 
     function buy() external payable {
         require(block.timestamp < expiresAt, "auction expired");
 
-        uint price = getPrice();
+        uint256 price = getPrice();
         require(msg.value >= price, "ETH < price");
 
         nft.transferFrom(seller, msg.sender, nftId);
-        uint refund = msg.value - price;
+        uint256 refund = msg.value - price;
         if (refund > 0) {
             payable(msg.sender).transfer(refund);
         }
