+++ src/pages/app/multi-call/MultiCall.sol
@@ -2,15 +2,16 @@
 pragma solidity ^0.8.20;
 
 contract MultiCall {
-    function multiCall(
-        address[] calldata targets,
-        bytes[] calldata data
-    ) external view returns (bytes[] memory) {
+    function multiCall(address[] calldata targets, bytes[] calldata data)
+        external
+        view
+        returns (bytes[] memory)
+    {
         require(targets.length == data.length, "target length != data length");
 
         bytes[] memory results = new bytes[](data.length);
 
-        for (uint i; i < targets.length; i++) {
+        for (uint256 i; i < targets.length; i++) {
             (bool success, bytes memory result) = targets[i].staticcall(data[i]);
             require(success, "call failed");
             results[i] = result;
