+++ src/pages/app/multi-sig-wallet/TestContract.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.20;
 
 contract TestContract {
-    uint public i;
+    uint256 public i;
 
-    function callMe(uint j) public {
+    function callMe(uint256 j) public {
         i += j;
     }
